+++ pochidetection/models/__init__.py
@@ -1,6 +1,7 @@
 """物体検出モデルパッケージ."""
 
 from pochidetection.models.rtdetr import RTDetrModel
+from pochidetection.models.ssd300 import SSD300Model
 from pochidetection.models.ssdlite import SSDLiteModel
 
-__all__ = ["RTDetrModel", "SSDLiteModel"]
+__all__ = ["RTDetrModel", "SSD300Model", "SSDLiteModel"]